feat: 중복 삭제 불가능 구현

+++ src/main/java/com/example/aboutme/service/SpaceService/SpaceServiceImpl.java
@@ -1,5 +1,7 @@
 package com.example.aboutme.service.SpaceService;
 
+import com.example.aboutme.apiPayload.code.status.ErrorStatus;
+import com.example.aboutme.apiPayload.exception.GeneralException;
 import com.example.aboutme.app.dto.SpaceRequest;
 import com.example.aboutme.converter.SpaceConverter;
 import com.example.aboutme.domain.Member;
@@ -7,10 +9,12 @@ import com.example.aboutme.domain.Space;
 import com.example.aboutme.repository.SpaceRepository;
 import com.example.aboutme.service.MemberService.MemberService;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 
+@Slf4j
 @Service
 @Transactional
 @AllArgsConstructor
@@ -26,14 +30,23 @@ public class SpaceServiceImpl implements SpaceService {
     @Override
     public Space readSpace(Long memberId) {
         Member member = memberService.findMember(memberId);
-        return spaceRepository.findByMember_Id(memberId);
+        // 존재하는지 검사
+        if (!spaceRepository.findByMember_Id(memberId).isPresent()) {
+            throw new GeneralException(ErrorStatus.SPACE_NOT_FOUND);
+        }
+        return spaceRepository.findByMember_Id(memberId).get();
     }
 
     @Override
     public Space deleteSpace(Long memberId) {
         Member member = memberService.findMember(memberId);
-        Space targetSpace = spaceRepository.findByMember_Id(memberId);
+        // 존재하는지 검사
+        if (!spaceRepository.findByMember_Id(memberId).isPresent()) {
+            throw new GeneralException(ErrorStatus.SPACE_NOT_FOUND);
+        }
+        Space targetSpace = spaceRepository.findByMember_Id(memberId).get();
         spaceRepository.delete(targetSpace);
+        log.info("여기까지 성공");
         return targetSpace;
     }
 }
